+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]
 ---
 
+### Added
+
+- `omegaconf` package to load YAML configuration files
+  - Defaults are loaded globally to avoid multiple files operations
+
 [1.8.0] - 2022-08-10
 ---